fix: move verify email to auth module
