+++ packages/wallet/backend/src/createContainer.ts
@@ -89,7 +89,9 @@ export const createContainer = (config: Env): Container<Bindings> => {
     async () =>
       new AuthService({
         env: await container.resolve('env'),
-        userService: await container.resolve('userService')
+        userService: await container.resolve('userService'),
+        logger: await container.resolve('logger'),
+        emailService: await container.resolve('emailService')
       })
   )
